pyproject.toml
[project]
name = "accident-response-chatbot"
version = "0.1.0"
requires-python = "==3.12.8"

dependencies = [
    "fastapi>=0.128,<0.130",
    "uvicorn[standard]>=0.40,<0.41",
    "langchain-openai>=1.1,<1.2",
    "langgraph>=1.0,<1.1",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.4.0",
    "pypdf>=6.6.0",
    "pymupdf>=1.26.7",
    "beautifulsoup4>=4.14.3",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "langchain-postgres>=0.0.16",
    "aiosqlite>=0.22.1",
    "ipykernel>=7.1.0",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
    "langchain-community>=0.4.1",
    "langchain>=1.2.7",
    "chromadb>=1.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
]

